FXG.3-d fix chain-adoption skipped by integer-type check
convertChainToGraphMode adopted 0 nodes, so converted effects fell back
to silence ("Effect is not active yet"). The bridge's napiToJson encodes
every JS number as a double, so the renderer-sent engineNodeId arrived as
3.0 (number_float); adoptGraphNodes gated on is_number_integer() and
skipped every entry. Accept any JSON number and cast to int. The native
test now sends engineNodeId as a double to cover the real bridge path.

+++ engine/test/test_effects.cpp
@@ -1135,8 +1135,11 @@ static void testGraphRuntimeOwnership()
         const int chainNode = chain.addEffect("testgain", 0);
         CHECK(chain.setEffectParameter(chainNode, "gain", 2.0f), "chain gain should be settable");
 
+        // engineNodeId is sent as a JSON double on purpose: the bridge's
+        // napiToJson encodes every JS number as a double, so adoptGraphNodes must
+        // accept number_float (not just number_integer) or it skips everything.
         const nlohmann::json mapping = nlohmann::json::array({
-            {{"effectInstanceId", "inst-adopt"}, {"engineNodeId", chainNode}},
+            {{"effectInstanceId", "inst-adopt"}, {"engineNodeId", static_cast<double>(chainNode)}},
         });
         const auto adoptRes = chain.adoptGraphNodes(mapping);
         CHECK(adoptRes.value("ok", false), "adoptGraphNodes should succeed for a live node");
